% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_function_Group 1.R
\name{coef_conf.int}
\alias{coef_conf.int}
\title{Confidence Interval for one linear combination}
\usage{
coef_conf.int(model, coef_test, conf.level = 0.95)
}
\arguments{
\item{model}{fitted model;}

\item{conf.level}{confidence level;}

\item{coef.test}{vector specifying linear combination of regression parameters for which a confidence interval is desired;}
}
\value{
Returns results of confidence interval for a single linear combination
}
\description{
Used to construct confidence interval interval for a single linear combination of parameters in a regression model.
}
\examples{
lm.fit <- lm(Sepal.Length ~Sepal.Width+Petal.Width, data=iris)
coef_conf.int(lm.fit,c(0,2,4),0.99)
}
