% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_function_Group 1.R
\name{f.test_coef}
\alias{f.test_coef}
\title{Hypothesis testing for several linear combinations}
\usage{
f.test_coef(model, coef_test, hypo_values)
}
\arguments{
\item{model}{fitted model;}

\item{hypo_values}{vector specifying hypothesized or null values;}

\item{coef.test}{matrix specifying linear combinations to test;}
}
\value{
Returns results of hypothesis testing for multiple linear combinations
}
\description{
Used to conduct an hypothesis testing of a one or more linear combinations of parameters in a regression model.
}
\examples{
lm.fit <- lm(Sepal.Length ~Sepal.Width+Petal.Width, data=iris)
my_matrix <- matrix(c(0,1,-1,-1,1,0), byrow = TRUE, nrow = 2)
my_hyp <- c(-0.6,-3)
f.test_coef(lm.fit, my_matrix, my_hyp)
}
