% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_function_Group 1.R
\name{t.test_coef}
\alias{t.test_coef}
\title{Hypothesis testing for one linear combination}
\usage{
\method{t}{test_coef}(model, coef_test, hypo_values)
}
\arguments{
\item{model}{fitted model;}

\item{hypo_values}{hypothesized or null value;}

\item{coef.test}{vector specifying linear combination to test;}
}
\value{
Returns results of hypothesis testing for a single linear combination
}
\description{
Used to conduct a hypothesis testa of a single linear combination of parameters in a regression model.
}
\examples{
lm.fit <- lm(Sepal.Length ~Sepal.Width+Petal.Width, data=iris)
t.test_coef(lm.fit,c(0,2,4),4)
}
